Add the module `convert_type` for converting different types of data
